requirements-linux.txt
# Linux-specific requirements for SQLite compatibility
pysqlite3-binary